HSEARCH-1001 Refactoring integration tests
